Splittet ut vilkårsperiode til egen service for å unngå circular dependency fra VilkårSteg og Stønadsperiode
