+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/StønadsperiodeService.kt
@@ -5,17 +5,19 @@ import no.nav.tilleggsstonader.sak.vilkår.domain.StønadsperiodeRepository
 import no.nav.tilleggsstonader.sak.vilkår.dto.StønadsperiodeDto
 import no.nav.tilleggsstonader.sak.vilkår.dto.tilSortertDto
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 import java.util.UUID
 
 @Service
 class StønadsperiodeService(
     private val stønadsperiodeRepository: StønadsperiodeRepository,
-    private val vilkårService: VilkårService,
+    private val vilkårperiodeService: VilkårperiodeService,
 ) {
     fun hentStønadsperioder(behandlingId: UUID): List<StønadsperiodeDto> {
         return stønadsperiodeRepository.findAllByBehandlingId(behandlingId).tilSortertDto()
     }
 
+    @Transactional
     fun lagreStønadsperioder(behandlingId: UUID, stønadsperioder: List<StønadsperiodeDto>): List<StønadsperiodeDto> {
         // TODO valider behandling er ikke låst
         validerStønadsperioder(behandlingId, stønadsperioder)
@@ -34,8 +36,13 @@ class StønadsperiodeService(
         ).tilSortertDto()
     }
 
-    fun validerStønadsperioder(behandlingId: UUID, stønadsperioder: List<StønadsperiodeDto>) {
-        val vilkårperioder = vilkårService.hentVilkårperioder(behandlingId)
+    fun validerStønadsperioder(behandlingId: UUID) {
+        val stønadsperioder = stønadsperiodeRepository.findAllByBehandlingId(behandlingId)
+        validerStønadsperioder(behandlingId, stønadsperioder.tilSortertDto())
+    }
+
+    private fun validerStønadsperioder(behandlingId: UUID, stønadsperioder: List<StønadsperiodeDto>) {
+        val vilkårperioder = vilkårperiodeService.hentVilkårperioder(behandlingId)
 
         StønadsperiodeValideringUtil.validerStønadsperioder(stønadsperioder, vilkårperioder)
     }
