+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/VilkårService.kt
@@ -9,34 +9,16 @@ import no.nav.tilleggsstonader.sak.infrastruktur.database.Sporbar
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.Feil
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvis
 import no.nav.tilleggsstonader.sak.opplysninger.søknad.SøknadService
-import no.nav.tilleggsstonader.sak.vilkår.domain.AktivitetType
-import no.nav.tilleggsstonader.sak.vilkår.domain.MålgruppeType
 import no.nav.tilleggsstonader.sak.vilkår.domain.Vilkår
 import no.nav.tilleggsstonader.sak.vilkår.domain.VilkårRepository
-import no.nav.tilleggsstonader.sak.vilkår.domain.VilkårType
-import no.nav.tilleggsstonader.sak.vilkår.domain.Vilkårperiode
 import no.nav.tilleggsstonader.sak.vilkår.domain.VilkårperiodeRepository
-import no.nav.tilleggsstonader.sak.vilkår.domain.VilkårperiodeType
-import no.nav.tilleggsstonader.sak.vilkår.dto.OpprettVilkårperiode
 import no.nav.tilleggsstonader.sak.vilkår.dto.VilkårDto
 import no.nav.tilleggsstonader.sak.vilkår.dto.VilkårGrunnlagDto
-import no.nav.tilleggsstonader.sak.vilkår.dto.VilkårperiodeDto
-import no.nav.tilleggsstonader.sak.vilkår.dto.Vilkårperioder
 import no.nav.tilleggsstonader.sak.vilkår.dto.VilkårsvurderingDto
 import no.nav.tilleggsstonader.sak.vilkår.dto.tilDto
 import no.nav.tilleggsstonader.sak.vilkår.regler.HovedregelMetadata
 import no.nav.tilleggsstonader.sak.vilkår.regler.evalutation.OppdaterVilkår
-import no.nav.tilleggsstonader.sak.vilkår.regler.evalutation.OppdaterVilkår.lagNyttVilkår
 import no.nav.tilleggsstonader.sak.vilkår.regler.evalutation.OppdaterVilkår.opprettNyeVilkår
-import no.nav.tilleggsstonader.sak.vilkår.regler.vilkår.AktivitetReelArbeidssøkerRegel
-import no.nav.tilleggsstonader.sak.vilkår.regler.vilkår.AktivitetTiltakRegel
-import no.nav.tilleggsstonader.sak.vilkår.regler.vilkår.AktivitetUtdanningRegel
-import no.nav.tilleggsstonader.sak.vilkår.regler.vilkår.MålgruppeAAPFerdigAvklartRegel
-import no.nav.tilleggsstonader.sak.vilkår.regler.vilkår.MålgruppeAAPRegel
-import no.nav.tilleggsstonader.sak.vilkår.regler.vilkår.MålgruppeDagpengerRegel
-import no.nav.tilleggsstonader.sak.vilkår.regler.vilkår.MålgruppeOmstillingsstønadRegel
-import no.nav.tilleggsstonader.sak.vilkår.regler.vilkår.MålgruppeOvergangsstønadRegel
-import no.nav.tilleggsstonader.sak.vilkår.regler.vilkår.MålgruppeUføretrygdRegel
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -79,62 +61,6 @@ class VilkårService(
         }
         return hentEllerOpprettVilkårsvurdering(behandlingId)
     }
-
-    fun hentVilkårperioder(behandlingId: UUID): Vilkårperioder {
-        val vilkår = vilkårRepository.findByBehandlingId(behandlingId)
-        val vilkårsperioder =
-            vilkårperiodeRepository.finnVilkårperioderForBehandling(behandlingId).associateBy { it.vilkårId }
-
-        return Vilkårperioder(
-            målgrupper = finnPerioder(vilkår, vilkårsperioder, VilkårType::gjelderMålgruppe),
-            aktiviteter = finnPerioder(vilkår, vilkårsperioder, VilkårType::gjelderAktivitet),
-        )
-    }
-
-    private fun finnPerioder(
-        vilkår: List<Vilkår>,
-        vilkårsperioder: Map<UUID, Vilkårperiode>,
-        gjelderFilter: (VilkårType) -> Boolean,
-    ) = vilkår.filter { gjelderFilter(it.type) }.map { vilkårsperioder.getValue(it.id).tilDto(it.tilDto()) }
-
-    @Transactional
-    fun opprettVilkårperiode(behandlingId: UUID, opprettVilkårperiode: OpprettVilkårperiode): VilkårperiodeDto {
-        feilHvis(behandlingErLåstForVidereRedigering(behandlingId)) {
-            "Kan ikke opprette vilkår når behandling er låst for videre redigering"
-        }
-
-        val vilkår = lagNyttVilkår(
-            vilkårsregel = vilkårsregelForVilkårsperiodeType(opprettVilkårperiode.type),
-            metadata = hentGrunnlagOgMetadata(behandlingId).second,
-            behandlingId = behandlingId,
-        )
-        vilkårRepository.insert(vilkår)
-
-        val vilkårperiode = vilkårperiodeRepository.insert(
-            Vilkårperiode(
-                vilkårId = vilkår.id,
-                fom = opprettVilkårperiode.fom,
-                tom = opprettVilkårperiode.tom,
-                type = opprettVilkårperiode.type,
-            ),
-        )
-
-        return vilkårperiode.tilDto(vilkår.tilDto())
-    }
-
-    private fun vilkårsregelForVilkårsperiodeType(vilkårperiodeType: VilkårperiodeType) =
-        when (vilkårperiodeType) {
-            MålgruppeType.AAP -> MålgruppeAAPRegel()
-            MålgruppeType.AAP_FERDIG_AVKLART -> MålgruppeAAPFerdigAvklartRegel()
-            MålgruppeType.DAGPENGER -> MålgruppeDagpengerRegel()
-            MålgruppeType.OMSTILLINGSSTØNAD -> MålgruppeOmstillingsstønadRegel()
-            MålgruppeType.OVERGANGSSTØNAD -> MålgruppeOvergangsstønadRegel()
-            MålgruppeType.UFØRETRYGD -> MålgruppeUføretrygdRegel()
-
-            AktivitetType.TILTAK -> AktivitetTiltakRegel()
-            AktivitetType.UTDANNING -> AktivitetUtdanningRegel()
-            AktivitetType.REEL_ARBEIDSSØKER -> AktivitetReelArbeidssøkerRegel()
-        }
 
     fun hentVilkårsett(behandlingId: UUID): List<VilkårDto> {
         val vilkårsett = vilkårRepository.findByBehandlingId(behandlingId)
